Implementing Collaborator Page View and anothers thinks.

+++ AgendaCCB.App/AgendaCCB.App/Models/UserAppSession.cs
@@ -5,8 +5,8 @@ namespace AgendaCCB.App.Models
     public class UserAppSession : RealmObject
     {
         [PrimaryKey]
-        public int RealmId { get; set; }
-        public int Id { get; set; }
+        public long RealmId { get; set; }
+        public long Id { get; set; }
         public string PhoneNumber { get; set; }
         public string Token { get; set; }
         public string Image { get; set; }
